+++ src/rex.restful/NEWS.rst
@@ -12,6 +12,7 @@ REX.RESTFUL Change Log
   setting the ``cors_policy`` property on the class that references a
   configuration defined by the ``restful_cors_policies`` application setting.
 * Added execution time to response logs.
+* Added support for the ``rex.web.Confine`` extension.
 
 
 1.1.0 (2016-01-29)